move blockwriter to another pkg
